Modified KikuchiPatternSimulator.on_detector() for detector Euler
Modified the calculation of the orientation matrix
u_s_bruker in KikuchiPatternSimulator.on_detector().
Previously u_s_bruker was determined from the 'total_tilt',
which only considered EBSDDetector.sample_tilt and
EBSDDetector.tilt but NOT EBSDDetector.azimuthal.
Now EBSDDetector has the detector Euler angles, which
describe the full orientation of the detector and these can
be used to calculate u_s_bruker. In this way, the full
orientation of the detector can be considered when making
simulations to overlay on patterns. This will give a much
better fit for cases where the detector azimuthal angle
was non-zero (i.e. EBSDDetector.euler[0] != 0) or where the
third detector Euler angle was non-zero.

Signed-off-by: tgwoodcock <tgwoodcock@users.noreply.github.com>